test: add Exceptions JVM serialization test
